% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Calculates the entropy}
\usage{
entropy(probs, freq = FALSE, base = "shannon")
}
\arguments{
\item{probs}{Probabilities to calculate entropy of}

\item{freq}{If set to \code{TRUE}, then \code{probs} are interpreted as frequencies. Defaults to FALSE.}

\item{base}{The base with respect to which logarithms are computed. Defaults to base 2="shannon". Possible options are "shannon" for base 2, "nat" for base e, and "hartley" for base 10.}
}
\description{
Calculates the entropy
}
