% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_to_right.R
\name{evaluate_left_to_right}
\alias{evaluate_left_to_right}
\title{Left-to-right evaluation algorithm}
\usage{
evaluate_left_to_right(test_documents, model, ntopics, alpha, beta, nparticles,
  resampling, seed = 0)
}
\description{
This is an algorithm for approximating p(w | ...) blabla
}
