% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor_mode.R
\name{get_bayes_factor_mode_from_data}
\alias{get_bayes_factor_mode_from_data}
\title{Creates a bayes factor mode given data}
\usage{
get_bayes_factor_mode_from_data(data)
}
\arguments{
\item{types}{A data frame of types and topic indicators.}
}
\description{
Creates a bayes factor mode given a data frame of types and topic indicators.
}
