% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor_mode.R
\name{read_bayes_factor_mode_from_file}
\alias{read_bayes_factor_mode_from_file}
\title{Reads a bayes factor mode from file}
\usage{
read_bayes_factor_mode_from_file(src)
}
\arguments{
\item{src}{A filename to where the mode data will be read.}
}
\description{
Reads a bayes factor mode from specified file location.
}
