% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor_mode.R
\name{type_to_topic_indicators}
\alias{type_to_topic_indicators}
\title{Type vector to topic indicator vector}
\usage{
type_to_topic_indicators(types, mode)
}
\arguments{
\item{types}{A vector of types.}

\item{mode}{A bayes factor mode.}
}
\description{
Translate a vector of types to their topic indicators given a mode.
}
