% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor_mode.R
\name{types_to_topic_indicators}
\alias{types_to_topic_indicators}
\title{Data frame of types to topic indicator vector}
\usage{
types_to_topic_indicators(types, mode)
}
\arguments{
\item{types}{A data frame containing the types}

\item{mode}{A bayes factor mode.}
}
\description{
Translate a data frame of types to their topic indicators given a mode.
}
