% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor_mode.R
\name{write_bayes_factor_mode_to_file}
\alias{write_bayes_factor_mode_to_file}
\title{Writes a bayes factor mode to file}
\usage{
write_bayes_factor_mode_to_file(mode, dst)
}
\arguments{
\item{mode}{A bayes factor mode.}

\item{dst}{A filename to where the mode data will be written.}
}
\description{
Writes a bayes factor mode to specified file location.
}
