% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_to_right.R
\name{evaluate_left_to_right}
\alias{evaluate_left_to_right}
\title{Left-to-right evaluation algorithm}
\usage{
evaluate_left_to_right(corpus, state, n_topics, alpha, beta, n_particles,
  resampling)
}
\description{
This is an algorithm for approximating p(w | ...) blabla
}
