% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factor_mode.R
\name{get_bayes_factor_mode_from_data}
\alias{get_bayes_factor_mode_from_data}
\title{Creates a bayes factor mode given data}
\usage{
get_bayes_factor_mode_from_data(types, topic_indicators)
}
\arguments{
\item{types}{A vector of types.}

\item{topic_indicators}{A vector of topic indicators.}
}
\description{
Creates a bayes factor mode given a type vector and a topic indicator vector.
}
